python-bridge/requirements.txt
firebase-admin==6.5.0
websockets==12.0
python-dotenv==1.0.1
